+++ priv/templates/lvn.gen/app_name_native.ex
@@ -125,7 +125,9 @@ defmodule <%= inspect context.native_module %> do
     quote do
       use LiveViewNative.Component, unquote(opts)
 
-      import LiveViewNative.Component, only: [csrf_token: 1]
+      # Import convenience functions from controllers
+      import Phoenix.Controller,
+        only: [get_csrf_token: 0, view_module: 1, view_template: 1]
 
       unquote(helpers(opts[:format]))
     end